backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Krzysztof Grabczyński <grabczynski.krzysztof@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"
pillow = "^10.3.0"
boto3 = "^1.34.143"
django-storages = "^1.14.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
